# src/tools/cleanroom-client/app/schema/openapi.yaml
openapi: 3.1.0
info:
  title: FastAPI
  description: Cleanroom Client API
  version: 0.1.0
paths:
...
  /login:
    post:
      summary: Login
      operationId: login_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /add-secretstore:
    post:
      summary: Add Secretstore
      operationId: add_secretstore_add_secretstore_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddSecretStoreRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /add-datastore:
    post:
      summary: Add Datastore
      operationId: add_datastore_add_datastore_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddDatastoreRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /upload-datastore:
    post:
      summary: Upload Datastore
      operationId: upload_datastore_upload_datastore_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadDataStoreRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /add-identity-az-federated:
    post:
      summary: Config Add Identity Az Federated
      operationId: config_add_identity_az_federated_add_identity_az_federated_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigAddIdentityAzFederatedRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /add-identity-az-secret:
    post:
      summary: Config Add Identity Az Secret
      operationId: config_add_identity_az_secret_add_identity_az_secret_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigAddIdentityAzSecretBased'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /add-identity-oidc-attested:
    post:
      summary: Config Add Identity Oidc Attested
      operationId: config_add_identity_oidc_attested_add_identity_oidc_attested_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigAddIdentityOIDCAttested'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /config/view:
    post:
      summary: Config Get
      operationId: config_get_config_view_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigViewRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            text/plain:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /config/init:
    post:
      summary: Config Init
      operationId: config_init_config_init_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigInitRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /config/validate:
    post:
      summary: Config Validate
      operationId: config_validate_config_validate_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigValidateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /config/add-datasource:
    post:
      summary: Config Add Datasource
      operationId: config_add_datasource_config_add_datasource_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigAddDatasourceRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /config/add-datasink:
    post:
      summary: Config Add Datasink
      operationId: config_add_datasink_config_add_datasink_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigAddDatasinkRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /datastore/download:
    post:
      summary: Datastore Download
      operationId: datastore_download_datastore_download_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataStoreDownloadRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /logs/download:
    post:
      summary: Logs Download
      operationId: logs_download_logs_download_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogsDownloadRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /telemetry/download:
    post:
      summary: Telemetry Download
      operationId: telemetry_download_telemetry_download_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetryDownloadRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /config/set-logging:
    post:
      summary: Config Set Logging
      operationId: config_set_logging_config_set_logging_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigSetLoggingRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /config/set-telemetry:
    post:
      summary: Config Set Telemetry
      operationId: config_set_telemetry_config_set_telemetry_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigSetTelemetryRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /config/add-application:
    post:
      summary: Config Add Application
      operationId: config_add_application_config_add_application_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigAddApplicationRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /config/add-application-endpoint:
    post:
      summary: Config Add Application Endpoint
      operationId: config_add_application_endpoint_config_add_application_endpoint_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigAddApplicationEndpointRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /config/create-kek:
    post:
      summary: Config Create Kek
      operationId: config_create_kek_config_create_kek_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigCreateKekRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /config/wrap-deks:
    post:
      summary: Config Wrap Deks
      operationId: config_wrap_deks_config_wrap_deks_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigWrapDeksRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /config/wrap-secret:
    post:
      summary: Config Wrap Secret
      operationId: config_wrap_secret_config_wrap_secret_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigWrapSecretRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AddDatastoreRequest:
      properties:
        configName:
          type: string
          title: Configname
        name:
          type: string
          title: Name
        secretStore:
          type: string
          title: Secretstore
        secretStoreConfig:
          type: string
          title: Secretstoreconfig
        encryptionMode:
          $ref: '#/components/schemas/EncryptionMode'
        backingStoreType:
          $ref: '#/components/schemas/BackingStoreType'
        backingStoreId:
          type: string
          title: Backingstoreid
        containerName:
          anyOf:
          - type: string
          - type: 'null'
          title: Containername
      type: object
      required:
      - configName
      - name
      - secretStore
      - secretStoreConfig
      - encryptionMode
      - backingStoreType
      - backingStoreId
      title: AddDatastoreRequest
    AddSecretStoreRequest:
      properties:
        name:
          type: string
          title: Name
        configName:
          type: string
          title: Configname
        backingStoreType:
          $ref: '#/components/schemas/SecretStoreType'
        backingStoreId:
          anyOf:
          - type: string
          - type: 'null'
          title: Backingstoreid
        backingStorePath:
          anyOf:
          - type: string
          - type: 'null'
          title: Backingstorepath
        attestationEndpoint:
          anyOf:
          - type: string
          - type: 'null'
          title: Attestationendpoint
      type: object
      required:
      - name
      - configName
      - backingStoreType
      title: AddSecretStoreRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LoginRequest:
      properties:
        loginArgs:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Loginargs
      type: object
      title: LoginRequest
...